Update README, minor Update print_sa function, rename quick.md to quick_commands.md
Update comes as attempt to run the script on a different PC, errors occured where module cgi was not found. This is due to The cgi module beign deprecated since Python 3.0, but it was still available in Python 3.x until Python 3.11. According to the Python documentation, the cgi module was removed in Python 3.12. Update to README to specify the Python version and requirement to revise code since googletrans library requires cgi